Updates remaining import paths according to renames
